+++ autoform-validation.js
@@ -1,66 +1,44 @@
-/* global _validateForm:true, AutoForm, validateField:true */
-
-/*
- * all form validation logic is here
- */
-
-_validateForm = function _validateForm(formId, formDoc, useCollectionSchema) {
-  var form = AutoForm.getCurrentDataForForm(formId);
-  var formType = form.type;
-
-  if (form.validation === 'none') {
-    return true;
-  }
-
-  // Call validateForm from the requested form type definition
-  var ftd = AutoForm._formTypeDefinitions[formType];
-  if (!ftd) {
-    throw new Error('AutoForm: Form type "' + formType + '" has not been defined');
-  }
-
-  return ftd.validateForm.call({
-    form: form,
-    formDoc: formDoc,
-    useCollectionSchema: useCollectionSchema
-  });
-};
+/* global AutoForm, validateField:true */
 
 function _validateField(key, formId, skipEmpty, onlyIfAlreadyInvalid) {
   var docToValidate;
 
   // Due to throttling, this can be called after the autoForm template is destroyed.
   // If that happens, we exit without error.
-  var template;
-  try{
-    template = AutoForm.templateInstanceForForm(formId);
-  }catch(e){}
+  var template = AutoForm.templateInstanceForForm(formId);
 
-  if (!template || !template.view._domrange || template.view.isDestroyed) {
-    return;
+  // If form is not currently rendered, return true
+  if (!template ||
+      !template.view ||
+      !template.view._domrange ||
+      template.view.isDestroyed) {
+    return true;
   }
 
   var form = AutoForm.getCurrentDataForForm(formId);
   var ss = AutoForm.getFormSchema(formId);
 
   if (!ss) {
-    return;
+    return true;
   }
 
   // Skip validation if onlyIfAlreadyInvalid is true and the form is
   // currently valid.
   if (onlyIfAlreadyInvalid && ss.namedContext(formId).isValid()) {
-    return; //skip validation
+    return true; //skip validation
   }
 
   // Create a document based on all the values of all the inputs on the form
   // Get the form type definition
-  var ftd = AutoForm._formTypeDefinitions[form.type];
-  if (!ftd) {
-    throw new Error('AutoForm: Form type "' + form.type + '" has not been defined');
-  }
+  var ftd = Utility.getFormTypeDef(form.type);
 
   // Clean and validate doc
   docToValidate = AutoForm.getFormValues(formId, template, ss, !!ftd.usesModifier);
+
+  // If form is not currently rendered, return true
+  if (!docToValidate) {
+    return true;
+  }
 
   // Skip validation if skipEmpty is true and the field we're validating
   // has no value.